backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Image processing
pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4

# AI/ML libraries
# PyTorch with CUDA support (Windows)
--index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1

# HuggingFace Transformers for diffusion models
transformers==4.46.0
diffusers==0.32.0
accelerate==1.0.0

# SAM2 for image segmentation
sam2 @ git+https://github.com/facebookresearch/segment-anything-2.git

# Additional ML utilities
safetensors==0.4.5
huggingface-hub==0.26.0

# Utilities
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.0
httpx==0.27.0
